Wait for memory changed
Before that I've been using a condition_variable and mutex pair in the main thread for instances when I don't have enough memory on the machine. That wasn't the best idea so I redid it. The current one isn't much better - just a while loop with some sleeps, so I think I'll come out with something better eventually.